+++ rumqttd/src/router/routing.rs
@@ -48,6 +48,8 @@ pub enum RouterError {
     UnsupportedQoS(QoS),
     #[error("Invalid filter prefix {0}")]
     InvalidFilterPrefix(Filter),
+    #[error("Invalid topic filter {0}: {1}")]
+    InvalidTopicFilter(Filter, String),
     #[error("Invalid client_id {0}")]
     InvalidClientId(String),
     #[error("Disconnection (Reason: {0:?})")]
@@ -554,9 +556,7 @@ impl Router {
                     self.router_meters.total_publishes += 1;
 
                     // Ignore retained messages
-                    if publish.retain {
-                        publish.retain = false;
-                    }
+                    publish.retain = false;
 
                     // Try to append publish to commitlog
                     match append_to_commitlog(
@@ -1416,8 +1416,59 @@ fn validate_subscription(
         return Err(RouterError::UnsupportedQoS(filter.qos));
     }
 
-    if filter.path.starts_with('$') {
-        return Err(RouterError::InvalidFilterPrefix(filter.path.to_owned()));
+    validate_filter_path(&filter.path)
+}
+
+/// Validate the filter path to ensure that it conforms to MQTT Spec Section 4.7
+fn validate_filter_path(filter_path: &str) -> Result<(), RouterError> {
+    trace!("Validating filter path = {}", filter_path);
+
+    if filter_path.is_empty() {
+        return Err(RouterError::InvalidTopicFilter(
+            filter_path.to_owned(),
+            "Topic filters must not be empty".to_owned(),
+        ));
+    }
+
+    if filter_path.contains("\0") {
+        return Err(RouterError::InvalidTopicFilter(
+            filter_path.to_owned(),
+            "Topic filters must not contain null bytes".to_owned(),
+        ));
+    }
+
+    // If we are seeing topics of 65kb or more it is almost certainly malicious
+    // and ideally we would reject those earlier than this but we certainly don't
+    // want to let them percolate further.
+    //
+    // The spec says that topics can not be any larger than this, but I don't see
+    // anything that prevents an implementation from choosing a smaller cap.
+    if filter_path.as_bytes().len() >= 65_536 {
+        return Err(RouterError::InvalidTopicFilter(
+            filter_path.chars().take(32).chain("...".chars()).collect(),
+            "Topic filters must be encoded in < 2^16 bytes".to_owned(),
+        ));
+    }
+
+    let mut levels = filter_path.split("/").peekable();
+
+    while let Some(level) = levels.next() {
+        if level == "+" {
+            continue;
+        } else if level == "#" {
+            if levels.peek() != None {
+                return Err(RouterError::InvalidTopicFilter(
+                    filter_path.to_owned(),
+                    "Topic filters may only use the multi-level wildcard # in the last level"
+                        .to_owned(),
+                ));
+            }
+        } else if level.contains("#") || level.contains("+") {
+            return Err(RouterError::InvalidTopicFilter(
+                filter_path.to_owned(),
+                "Topic filter wild cards must not be mixed with string literals".to_owned(),
+            ));
+        }
     }
 
     Ok(())
@@ -1431,6 +1482,57 @@ fn validate_clientid(client_id: &str) -> Result<(), RouterError> {
     }
 
     Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use assert_matches::assert_matches;
+
+    #[test]
+    fn test_validate_filter_path() {
+        assert_matches!(
+            validate_filter_path(""),
+            Err(RouterError::InvalidTopicFilter(_, _))
+        );
+        assert_matches!(validate_filter_path("/"), Ok(_));
+        assert_matches!(validate_filter_path("//"), Ok(_));
+        assert_matches!(validate_filter_path("#"), Ok(_));
+        assert_matches!(validate_filter_path("+"), Ok(_));
+        assert_matches!(validate_filter_path("sport/tennis/player1/#"), Ok(_));
+        assert_matches!(
+            validate_filter_path("sport/tennis#"),
+            Err(RouterError::InvalidTopicFilter(_, _))
+        );
+        assert_matches!(
+            validate_filter_path("sport/tennis/#/ranking"),
+            Err(RouterError::InvalidTopicFilter(_, _))
+        );
+        assert_matches!(validate_filter_path("+/tennis/#"), Ok(_));
+        assert_matches!(
+            validate_filter_path("sport+"),
+            Err(RouterError::InvalidTopicFilter(_, _))
+        );
+        assert_matches!(validate_filter_path("sport/+/player1"), Ok(_));
+        assert_matches!(validate_filter_path("$SYS/#"), Ok(_));
+        assert_matches!(validate_filter_path("accounts payable"), Ok(_));
+        assert_matches!(
+            validate_filter_path("hello/w\u{0000}rld"),
+            Err(RouterError::InvalidTopicFilter(_, _))
+        );
+    }
+    #[test]
+    fn test_validate_filter_path_too_large() {
+        let mut malicious_filter: String = "hello/world/".chars().cycle().take(65535).collect();
+        assert_matches!(validate_filter_path(malicious_filter.as_str()), Ok(_));
+
+        malicious_filter.push_str("hello");
+        assert_matches!(
+            validate_filter_path(malicious_filter.as_str()),
+            Err(RouterError::InvalidTopicFilter(_, _))
+        );
+    }
 }
 
 // #[cfg(test)]